Split SampleStrategy to a separate class file.
